integrating generalized json loader, optimizing input system & fmod audio hotpaths

+++ Source/Nomad.Audio.FMOD/Public/ValueObjects/FMODBankLoadingStrategy.cs
@@ -15,7 +15,7 @@ of merchantability, fitness for a particular purpose and noninfringement.
 
 namespace Nomad.Audio.Fmod.ValueObjects
 {
-    public enum FMODBankLoadingStrategy : uint
+    public enum FMODBankLoadingStrategy : byte
     {
         Streaming,
         Decompressed,
